requirements.txt
Django==3.2.16
django-environ==0.9.0
djangorestframework==3.14.0
django-cors-headers==3.13.0
kubernetes==25.3.0